+++ Zahra/src/Platform/Vulkan/VulkanShaderResourceManager.cpp
@@ -68,7 +68,7 @@ namespace Zahra
 
 		VkDescriptorPoolCreateInfo poolInfo{};
 		poolInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-		poolInfo.poolSizeCount = poolSizes.size();
+		poolInfo.poolSizeCount = (uint32_t)poolSizes.size();
 		poolInfo.pPoolSizes = poolSizes.data();
 		poolInfo.maxSets = (maxSetIndex + 1) * framesInFlight;
