+++ Zahra/src/Zahra/Renderer/IndexBuffer.cpp
@@ -6,7 +6,7 @@
 
 namespace Zahra
 {
-	Ref<IndexBuffer> IndexBuffer::Create(const uint32_t* indices, uint64_t count)
+	Ref<IndexBuffer> IndexBuffer::Create(const uint32_t* indices, uint32_t count)
 	{
 		switch (Renderer::GetAPI())
 		{
